Refactored aes-ni.* files to be under src/crypto
